+++ data_test.go
@@ -12,7 +12,7 @@ import (
 
 func TestDataSetGet(t *testing.T) {
 	var d Data
-	for _, want := range []time.Time{time.Now(), time.Time{}} {
+	for _, want := range []time.Time{time.Now(), {}} {
 		d.SetTime(want)
 		if got := d.GetTime(); !got.Equal(want) && got.Format(time.RFC3339) != want.Format(time.RFC3339) {
 			t.Errorf("set %v, got %v", want, got)